pagination servlet ok

+++ CDBwebui/src/main/java/DTO/CompanyDTO.java
@@ -2,22 +2,22 @@ package DTO;
 
 public class CompanyDTO {
 
-	private Long id;
+	private int id;
 	private String name;
 	
-	public CompanyDTO(long id) {
+	public CompanyDTO(int id) {
 		this.id=id;
 	}
 	public CompanyDTO() {
 		super();
 	}
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 	public String getName() {
 		return name;
 	}
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public void setName(String name) {
